A lot of small design changes all at once.
Why:

Breaking out dataset, project, and tool partials.

Changing all the green (.positive) buttons to blue (.primary).

Changing the service block padding-right to 4rem instead of 5, because 5
caused longer text to wrap weirdly.

Adding and implementing a helper so that, in the dataset partial, we see
either the first sentence or the first 140 characters, whichever is
shorter.

Adding a preview of the current thumbnail to the form while editing,
because it was unclear whether the resource had no image, or we just
weren't seeing it.

Changing routes so we don't have dataset/show at any point -- there's no
view for it.

Changing all link_to 'mailto:X' to mail_to links, and changing many %a
tags to link_to's.

Changing the "Get to know us." link to "Meet our team."

Automatic removal of trailing whitespace in a lot of places.